% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{merge_geoproc_vo}
\alias{merge_geoproc_vo}
\title{Title
merges geoproc with volcano file  of any file}
\usage{
merge_geoproc_vo(df, v, sub = T)
}
\arguments{
\item{sub}{}
}
\value{

}
\description{
Title
merges geoproc with volcano file  of any file
}
